pipeline-azure/requirements.txt
azure-functions
azure-storage-blob
scikit-learn
pandas
pyarrow
joblib